First step using python turn_events for tutorial sitreps.
Move triggering the tutorial sitrep about building ships from the C++
code, where it didn't belong, to the python/turn_events module
as suggested by @Vezzra and @geoffthemedio.  This also fixes a
problem with the old implementation: it sometimes asked the user
to start building ships even though no Basic Shipyard was available
yet.

This patch includes exposing Empire::ResearchedTechs to python because
Empire::TechResearched doesn't provide information about the turn of
discovery, and it adds "label" support to the generate_sitrep() method:
That method wasn't used yet, so changing the API shouldn't cause any
trouble, and i think specifying the label will be required for
practically every sitrep.  @Vezzra explicitly stated that the API
is intended to be enhanced as required in practice.

The other sitrep triggers will have to be moved in subsequent commits.