Trigger all tutorial hints from turn_events.py.
All but one that exist so far require it anyway due to the inspections and
calculations that need to be done, and having them all in turn_events.py
allows depending on the starting era in a uniform manner.

In pre-warp mode, give the player a Scout ship on turn 2.  Letting the
players research and build everything needed by themselves would
slow down the game too much.

While here, for SITREP_TUTORIAL_FOCUS, avoid showing it during the
same turn as some other tutorial step.